* Many documentation updates have been made. Note that the documentation has been relicensed under the MIT license, instead of the GNU FDL.
* Many facets of VMIPS are now configurable at runtime, instead of at compile time. For instance:
	It is now possible to build and install VMIPS without having previously installed MIPS cross-compiler tools. In particular, VMIPS now incorporates the portions of GNU libopcodes used to implement the MIPS disassembler, so linking against an installed version of libopcodes from GNU binutils is no longer necessary.
	It is now possible to switch the VMIPS CPU from big-endian to little-endian mode or vice-versa using a command-line option.
	It is now possible to change the name of the configuration file that VMIPS reads on startup using a command-line option. The "-o configfile" option, which never worked, has been removed.

* vmipstool now shares command-line processing code with vmips. It also has a new -swap-words option that allows you to byteswap all the 4-byte words in a file.

* You can now specify numeric constants in command line options which are multiples of 1024, 1024^2, or 1024^3 using the K, M, or G suffixes.

* Several new emulated devices based on the DECstation 5000/200 have been added.  In particular, it is no longer possible to disable the serial device at compile time (or with `-o nousetty'), but you can choose whether to use the DECstation serial chip or the SPIM-compatible serial console (or neither) using new command-line options. In addition, the ROM program distributed with VMIPS supports a "boot environment" similar to that provided by the DECstation PROM. The new DECstation-compatible devices have not been comprehensively validated; they should be considered "beta" quality at this point.

* There is a new tracing framework you can use to generate runtime execution traces of programs that run in VMIPS. See the documentation of the "tracing" command-line option and the other options that begin with "trace" for more details.

* The VMIPS interface to GDB now supports the "remote Z-packet" interface for setting breakpoints.

* VMIPS now prints more informative diagnostic messages about the causes of bus errors and interrupts.

* You can now use ^C (your terminal's Interrupt key) to stop VMIPS and break into the debugger, if the `-o debug' command line option was set, or ^\ (your terminal's Quit key) to halt VMIPS in an orderly manner at any time.

* The `--disable-debug' option to `configure' now strips the compiled VMIPS binaries, although it has been documented to do this for some time now.

+++ sample_code/xmboot/xmrcv.c
@@ -1,6 +1,6 @@
 #include "xmrcv.h"
 
-int xmrcv(int *bytes_rcvd, unsigned char *buf)
+int xmrcv(long *bytes_rcvd, unsigned char *buf)
 {
 	int err = 0, x, i, packetno = 0;
 	unsigned char ch, seq, cmpl_seq, cksum, their_cksum;
